Add files via upload
There are five main components to the IOTA-IF-Certify code:

The IOTA Software developments Kit (SDK): - This is the low level basic Library that the next layer up (Identity is built upon). It is the used to help design the systems that work on the IOTA Framework. The IOTA SDK kit is provided by the IOTA development team.

IOTA Identity: - Is the Library that implements the functionality that is specified in the 'W3C DID specification'. It implements that specification using the IOTA SDK above.T  Again the IOTA identity Library is provided by the IOTA development team and both are dependencies of the IOTA certify programming below. 

IOTA Certify: - Is the Library of functions that are more defined in what is required within the 'index.ts'. This Layer is the convenience layer that accommodates the specified changes to the script within the 'index.ts' providing the user interactions. Generating the commands to create the identity for Hans