Add files via upload
IOTA-IF-Certify 
For the purposes of the Proof of Concept (POC) it is important to note that the required security standards required for the safe storage of secret identity tokens and personal data will not be held in a secure onsite / cloud environment as to do so would require the use of the MainNet and the purchase of the crypto Currency (MIOTA) - The TestNet will suffice to demonstrate same with the use of the Faucet. 

There are five main components to the IOTA-IF-Certify code:

The IOTA Software developments Kit (SDK) Layer: - This is the low level basic Library that the next layer up (Identity is built upon). It is the used to help design the systems that work on the IOTA Framework. The IOTA SDK kit is provided by the IOTA development team.

IOTA Identity Layer: - Is the Library that implements the functionality that is specified in the 'W3C DID specification'. It implements that specification using the IOTA SDK above.T  Again the IOTA identity Library is provided by the IOTA development team and both are dependencies of the IOTA certify programming below. 

IOTA Certify Layer: - Is the Library of functions that are more defined in what is required within the 'index.ts'. This Layer is the convenience layer that accommodates the specified changes to the script within the 'index.ts' providing the user interactions. Generating the commands to create the identity for Hans (The Holder), the College, (The Issuer), & the CRU, (The Verifier).

The Application Layer: - In this Proof of Concept (PoC) the provision of tokens, which is required to store the Hans data on the Network and to validate the tasks are therefore allocated via an abstracted away utility function via the TestNet. The POC Script file (POC.ts) facilitates these requests.

Proof of Concept Layer: - When required to create tokens within the normal utility functionality in a production environment of the application layer of the MainNet, this would be provided via the Application Layer to work and output the tokens and validation etc.. The POC script provided will do that facility within the POC.ts file.

Other portions of the IOTA-IF-Certify POC:

Node Network: - the IOTA SDK interacts with the entire Node Network, which in this PoC is made up of Hans (The Holder), the College, (The Issuer), & the CRU, (The Verifier). In the production environment these woudl all be individual accounts on local storage or in the Cloud, however the principals observed in this PoC witl still demonstrate that functionality.

IOTA Tangle: - is a Directed Acyclic Graph (DAG) consensus algorithm, which eliminates the need for miners or validators and blocks. It allows for feeless, scalable, and near-instant transactions, making it suitable for the IoT environment. When a user sends an IOTA transaction, they validate two other transactions, enabling the network to overcome the cost and scalability limitations of blockchain. The IOTA Tangle is the backbone of IOTA's network, immutably recording the exchange of data and value, ensuring trustworthy information that cannot be tampered with or destroyed2.

Note: In this POC the data is only stored to memory and not to a local drive / disk.